# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telephone Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Call {
  id: number;
  callSid: string;
  fromNumber: string;
  toNumber: string;
  status: string;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export type CallStatsCallsByStatus = {[key: string]: number};

export interface CallStats {
  totalCalls: number;
  callsToday: number;
  /** @nullable */
  averageDurationSeconds: number | null;
  callsByStatus: CallStatsCallsByStatus;
}

export interface AgentConfig {
  id: number;
  systemPrompt: string;
  voice: string;
  greeting: string;
  updatedAt: string;
}

export interface AgentConfigInput {
  systemPrompt?: string;
  voice?: string;
  greeting?: string;
}

export type ListCallsParams = {
limit?: number;
offset?: number;
};