# lib/api-zod/src/generated/types/agentConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telephone Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentConfig {
  id: number;
  systemPrompt: string;
  voice: string;
  greeting: string;
  updatedAt: Date;
}