# lib/api-zod/src/generated/types/agentConfigInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telephone Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentConfigInput {
  systemPrompt?: string;
  voice?: string;
  greeting?: string;
}