# lib/api-zod/src/generated/types/call.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telephone Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Call {
  id: number;
  callSid: string;
  fromNumber: string;
  toNumber: string;
  status: string;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  summary?: string | null;
  createdAt: Date;
  /** @nullable */
  endedAt?: Date | null;
}