# lib/api-zod/src/generated/types/callStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Telephone Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { CallStatsCallsByStatus } from './callStatsCallsByStatus';

export interface CallStats {
  totalCalls: number;
  callsToday: number;
  /** @nullable */
  averageDurationSeconds: number | null;
  callsByStatus: CallStatsCallsByStatus;
}